+++ cargo-smart-release/tests/snapshots/triple-depth-workspace-changelog/a-dry-run-success-multi-crate
@@ -1,0 +1,1 @@
+[INFO ] WOULD write 2 sections to a/CHANGELOG.md
